Indent Changes changes done by Thangaraj

+++ openerp-server/openerp/addons/kg_brandmoc_rate/kg_brandmoc_rate.py
@@ -8,14 +8,10 @@ import re
 import math
 from datetime import date
 
-dt_time = time.strftime('%m/%d/%Y %H:%M:%S')
-
 class kg_brandmoc_rate(osv.osv):
 
 	_name = "kg.brandmoc.rate"
 	_description = "Brand Moc Rate Master"
-	
-	
 	
 	def _get_modify(self, cr, uid, ids, field_name, arg,  context=None):
 		res={}		
@@ -65,7 +61,7 @@ class kg_brandmoc_rate(osv.osv):
 		
 		
 		'modify': fields.function(_get_modify, string='Modify', method=True, type='char', size=10),	
-		'latest_price': fields.related('product_id','latest_price', type='float', string='Latest Price(Rs)', store=True),
+		'latest_price': fields.related('product_id','latest_price', type='float', string='Latest Price(Rs)', store=True,readonly=True),
 		'category_type': fields.selection([('purchase_item','Purchase Item'),('design_item','Design Item')],'Category'),
 		
 		
@@ -74,12 +70,12 @@ class kg_brandmoc_rate(osv.osv):
 		'copy_flag':fields.boolean('Copy Flag'),
 		
 		### Entry Info ###
-		'crt_date': fields.datetime('Creation Date',readonly=True),
+		'crt_date': fields.datetime('Created Date',readonly=True),
 		'user_id': fields.many2one('res.users', 'Created By', readonly=True),
 		'confirm_date': fields.datetime('Confirmed Date', readonly=True),
 		'confirm_user_id': fields.many2one('res.users', 'Confirmed By', readonly=True),
-		'ap_rej_date': fields.datetime('Approved/Reject Date', readonly=True),
-		'ap_rej_user_id': fields.many2one('res.users', 'Approved/Reject By', readonly=True),	
+		'ap_rej_date': fields.datetime('Approved/Rejected Date', readonly=True),
+		'ap_rej_user_id': fields.many2one('res.users', 'Approved/Rejected By', readonly=True),	
 		'cancel_date': fields.datetime('Cancelled Date', readonly=True),
 		'cancel_user_id': fields.many2one('res.users', 'Cancelled By', readonly=True),
 		'update_date': fields.datetime('Last Updated Date', readonly=True),
@@ -120,24 +116,25 @@ class kg_brandmoc_rate(osv.osv):
 		
 	def entry_cancel(self,cr,uid,ids,context=None):
 		rec = self.browse(cr,uid,ids[0])
-		if rec.product_id:
-			for item in rec.line_ids:
-				brand_rec = self.pool.get('kg.brand.master').browse(cr,uid,item.brand_id.id)				
-				sql_check = """ select brd_id,prod_id from prod_brnd where brd_id=%s and prod_id=%s""" %(brand_rec.id,rec.product_id.id)
-				cr.execute(sql_check)
-				data = cr.dictfetchall()
-				
-				if data:
-					sql = """ delete from prod_brnd where brd_id =%s and prod_id=%s """ %(brand_rec.id,rec.product_id.id)
-					cr.execute(sql)
-				else:				
-					#~ self.pool.get('kg.brand.master').write(cr,uid,brand_rec.id,{'product_ids':[(6, 0, [rec.product_id.id])]})				
-					pass
-		if rec.cancel_remark:
-			self.write(cr, uid, ids, {'state': 'cancel','cancel_user_id': uid, 'cancel_date': time.strftime('%Y-%m-%d %H:%M:%S')})
-		else:
-			raise osv.except_osv(_('Cancel remark is must !!'),
-				_('Enter the remarks in Cancel remarks field !!'))
+		if rec.state == 'approved':
+			if rec.product_id:
+				for item in rec.line_ids:
+					brand_rec = self.pool.get('kg.brand.master').browse(cr,uid,item.brand_id.id)				
+					sql_check = """ select brd_id,prod_id from prod_brnd where brd_id=%s and prod_id=%s""" %(brand_rec.id,rec.product_id.id)
+					cr.execute(sql_check)
+					data = cr.dictfetchall()
+					
+					if data:
+						sql = """ delete from prod_brnd where brd_id =%s and prod_id=%s """ %(brand_rec.id,rec.product_id.id)
+						cr.execute(sql)
+					else:				
+						#~ self.pool.get('kg.brand.master').write(cr,uid,brand_rec.id,{'product_ids':[(6, 0, [rec.product_id.id])]})				
+						pass
+			if rec.cancel_remark:
+				self.write(cr, uid, ids, {'state': 'cancel','cancel_user_id': uid, 'cancel_date': time.strftime('%Y-%m-%d %H:%M:%S')})
+			else:
+				raise osv.except_osv(_('Cancel remark is must !!'),
+					_('Enter the remarks in Cancel remarks field !!'))
 		return True
 		
 	def onchange_product(self, cr, uid, ids, product_id, context=None):		
@@ -163,7 +160,9 @@ class kg_brandmoc_rate(osv.osv):
 		return True
 		
 	def entry_draft(self,cr,uid,ids,context=None):
-		self.write(cr, uid, ids, {'state': 'draft'})
+		rec = self.browse(cr,uid,ids[0])
+		if rec.state == 'approved':
+			self.write(cr, uid, ids, {'state': 'draft'})
 		return True
 
 	def entry_confirm(self,cr,uid,ids,context=None):
